GPU/cndb/requirements.txt
sqlalchemy==1.3.23
sqlalchemy_utils==0.36.8
psycopg2-binary>=2.8.6
pyyaml>=1.15.0
